requirements.txt
python-dotenv~=1.0.1
akeyless~=4.2.5